treewide: specify field width as int
Format strings expect the field width (if given) to be specified by an
int.  The compiler will complain when compiling with
"-Wformat-signedness" otherwise.  Cast values that are not ints to int
when using them to specify the field length.

This is safe to do.  The formatting functions already expect an int
here, and if there is an overflow, this is a bug anyway which we
should fix separately.

+++ builtin/ls-files.c
@@ -167,7 +167,7 @@ static void show_killed_files(const struct index_state *istate,
 				pos = index_name_pos(istate, ent->name, ent->len);
 				if (0 <= pos)
 					BUG("killed-file %.*s not found",
-						ent->len, ent->name);
+					    (int)ent->len, ent->name);
 				pos = -pos - 1;
 				while (pos < istate->cache_nr &&
 				       ce_stage(istate->cache[pos]))
