+++ builtin/name-rev.c
@@ -389,7 +389,7 @@ static void name_rev_line(char *p, struct name_ref_data *data)
 				continue;
 
 			if (data->name_only)
-				printf("%.*s%s", p_len - hexsz, p_start, name);
+				printf("%.*s%s", (int)(p_len - hexsz), p_start, name);
 			else
 				printf("%.*s (%s)", p_len, p_start, name);
 			p_start = p + 1;
